# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OG Miner: OpenGraph Miner",
    "description": "High Performance OpenGraph Miner backed by FastAPI and Redis. Supports Headless Browser execution for SPAs, SSRF protection, and rich data extraction including OpenGraph, Twitter Cards, and JSON-LD.",
    "contact": {
      "name": "Kavindu Gunasena",
      "url": "https://www.linkedin.com/in/sasmithakg/"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/v1/extract": {
      "post": {
        "tags": [
          "extract"
        ],
        "summary": "Extract Metadata",
        "operationId": "extract_metadata",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/screenshot": {
      "get": {
        "tags": [
          "screenshot"
        ],
        "summary": "Take Screenshot",
        "description": "Capture a screenshot of a website.\nReturns a PNG image directly.",
        "operationId": "take_screenshot",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 2083,
              "title": "Url"
            }
          },
          {
            "name": "full_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Full Page"
            }
          },
          {
            "name": "width",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3840,
              "minimum": 320,
              "default": 1280,
              "title": "Width"
            }
          },
          {
            "name": "height",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2160,
              "minimum": 240,
              "default": 720,
              "title": "Height"
            }
          },
          {
            "name": "delay",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Delay"
            }
          },
          {
            "name": "dark_mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Dark Mode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/batch/extract": {
      "post": {
        "tags": [
          "batch"
        ],
        "summary": "Batch Extraction",
        "description": "Extract metadata for multiple URLs in parallel.",
        "operationId": "batch_extract",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchExtractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/image": {
      "get": {
        "tags": [
          "image"
        ],
        "summary": "Image Proxy",
        "description": "Proxies and optionally resizes an image. \nResults are cached for performance.\nReturns WebP format by default.",
        "operationId": "image_proxy",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The URL of the image to proxy",
              "title": "Url"
            },
            "description": "The URL of the image to proxy"
          },
          {
            "name": "width",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2000,
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Target width",
              "title": "Width"
            },
            "description": "Target width"
          },
          {
            "name": "height",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2000,
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Target height",
              "title": "Height"
            },
            "description": "Target height"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BatchExtractRequest": {
        "properties": {
          "urls": {
            "items": {
              "type": "string",
              "maxLength": 2083,
              "minLength": 1,
              "format": "uri"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Urls",
            "description": "List of URLs to extract metadata from (max 50)."
          },
          "enable_javascript": {
            "type": "boolean",
            "title": "Enable Javascript",
            "description": "Use headless browser for all URLs.",
            "default": false
          },
          "force_refresh": {
            "type": "boolean",
            "title": "Force Refresh",
            "description": "Bypass cache for all URLs.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "urls"
        ],
        "title": "BatchExtractRequest",
        "example": {
          "enable_javascript": false,
          "urls": [
            "https://google.com",
            "https://apple.com"
          ]
        }
      },
      "BatchExtractResponse": {
        "properties": {
          "results": {
            "additionalProperties": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExtractResponse"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "object",
            "title": "Results",
            "description": "Map of URL to extraction result or error."
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "successful": {
            "type": "integer",
            "title": "Successful"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          }
        },
        "type": "object",
        "required": [
          "results",
          "total",
          "successful",
          "failed"
        ],
        "title": "BatchExtractResponse"
      },
      "ExtractRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "force_refresh": {
            "type": "boolean",
            "title": "Force Refresh",
            "default": false
          },
          "enable_javascript": {
            "type": "boolean",
            "title": "Enable Javascript",
            "default": false
          },
          "proxy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "cookies": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cookies"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ExtractRequest",
        "example": {
          "country": "US",
          "enable_javascript": false,
          "url": "https://github.com"
        }
      },
      "ExtractResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          },
          "data": {
            "$ref": "#/components/schemas/Metadata"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "ExtractResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MetaInfo": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "url",
          "domain",
          "latency_ms"
        ],
        "title": "MetaInfo"
      },
      "Metadata": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "favicon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favicon"
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "site_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Name"
          },
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "oembed": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oembed"
          },
          "json_ld": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Json Ld"
          }
        },
        "type": "object",
        "title": "Metadata"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Proxy-Secret"
      }
    }
  }
}